// src/lib/api/services/QuotesService/QuotesService.ts
export interface CurrencySymbol {
  currencyCode: string;
  decimal: string;
  description: string;
  id: number;
  precision: number;
  symbol: string;
  tenantId: number;
  thousand: string;
}

// Approval Group Interface
export interface ApprovalGroup {
  approvalName: string;
  discount: boolean;
  id: number;
  tenantId: number;
}

// Quote User Interface
export interface QuoteUser {
  createdOn: string | null;
  department: string | null;
  displayName: string;
  dottedReportingTo: string | null;
  email: string;
  id: number;
  isActive: boolean | null;
  isDeleted: boolean | null;
  jobTitle: string | null;
  lastLoginAt: string | null;
  mobileNo: string | null;
  name: string | null;
  nationalMobileNum: string | null;
  ownerRole: string | null;
  password: string | null;
  picture: string;
  reportingTo: string | null;
  roleId: number | null;
  roleName: string | null;
  status: string;
  updatedOn: string | null;
  userBranchBUDTO: unknown | null;
  userCode: string | null;
  userTags: string | null;
  vendorId: number | null;
}

// Quote Item Interface
export interface QuoteItem {
  SPRRequested: boolean;
  approvalAwaiting: unknown | null;
  approvalGroupId: ApprovalGroup | null;
  approvalInitiated: boolean;
  buyerCompanyBranchName: string;
  buyerCompanyName: string;
  calculatedTotal: number;
  createdDate: string;
  curencySymbol: CurrencySymbol;
  customerRequiredDate: string | null;
  erpCode: string | null;
  grandTotal: number;
  itemCount: number;
  lastUpdatedDate: string;
  leadIdentifier: string | null;
  orderIdentifier: string[] | null;
  quotationIdentifier: string;
  quoteName: string;
  quoteUsers: QuoteUser[] | null;
  sellerCompanyBranchName: string;
  sellerCompanyName: string;
  subTotal: number;
  taxableAmount: number;
  updatedBuyerStatus: string;
  updatedSellerStatus: string;
  vendorId: number | null;
}